+++ lib/src/infra/adapters/micro_app_event/micro_app_event_adapter.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/services.dart';
+
+import '../../../entities/events/micro_app_event.dart';
+
+abstract class MicroAppEventAdapter {
+  MicroAppEvent parse(MethodCall methodCall);
+}